+++ src/main/java/com/github/unchama/seichiassist/data/button/PlayerDataButtons.java
@@ -3,15 +3,14 @@ package com.github.unchama.seichiassist.data.button;
 import com.github.unchama.seichiassist.SeichiAssist;
 import com.github.unchama.seichiassist.data.PlayerData;
 import com.github.unchama.seichiassist.data.descrptions.PlayerInfomationDescriptions;
-import com.github.unchama.seichiassist.data.itemstack.builder.SkullItemStackBuilder;
-import com.github.unchama.seichiassist.data.slot.button.Button;
-import com.github.unchama.seichiassist.data.slot.button.ButtonBuilder;
-import com.github.unchama.seichiassist.data.slot.handler.SlotActionHandler;
-import com.github.unchama.seichiassist.data.slot.handler.Triggers;
+import com.github.unchama.seichiassist.data.inventory.itemstack.builder.SkullItemStackBuilder;
+import com.github.unchama.seichiassist.data.inventory.slot.button.Button;
+import com.github.unchama.seichiassist.data.inventory.slot.button.ButtonBuilder;
+import com.github.unchama.seichiassist.data.inventory.slot.handler.SlotActionHandler;
+import com.github.unchama.seichiassist.data.inventory.slot.handler.Triggers;
 import com.github.unchama.seichiassist.text.Text;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
-import org.bukkit.event.player.PlayerInteractEvent;
 
 /**
  * @author karayuu
