+++ src/main/java/com/github/unchama/seichiassist/data/player/MiningSpeed.java
@@ -1,0 +1,10 @@
+package com.github.unchama.seichiassist.data.player;
+
+/**
+ * プレイヤーの採掘速度に関するデータ
+ *
+ * Created by karayuu on 2019/05/06
+ */
+public class MiningSpeed {
+
+}
