--- src/main/java/com/github/unchama/seichiassist/data/slot/handler/SlotActionHandlers.java
@@ -1,17 +1,0 @@
-package com.github.unchama.seichiassist.data.slot.handler;
-
-import com.github.unchama.seichiassist.data.slot.button.Button;
-
-/**
- * @author karayuu
- */
-public class SlotActionHandlers {
-    /**
-     * {@link Button} をReadOnly(移動不可)に可能な {@link SlotActionHandler} です.
-     * {@link Button} には常に付与されます.
-     */
-    public static SlotActionHandler READ_ONLY = new SlotActionHandler(
-            event -> true,
-            event -> event.setCancelled(true)
-    );
-}
