+++ src/main/java/com/github/unchama/seichiassist/listener/PlayerClickListener.java
@@ -526,12 +526,12 @@ public class PlayerClickListener implements Listener {
 				}
 				//開く音を再生
 				player.playSound(player.getLocation(), Sound.BLOCK_FENCE_GATE_OPEN, 1, (float) 0.1);
-				Inventory inv = MenuInventoryData.getMenuData(player);
+				/*Inventory inv = MenuInventoryData.getMenuData(player);
 				if(inv == null){
 					return;
 				}
-				player.openInventory(inv);
-				//StickMenu.stickMenu.open(player);
+				player.openInventory(inv);*/
+				StickMenu.stickMenu.open(player);
 			}
 		}
 	}
